Create apiRequest.R
Copied over code from the WS repo, and also changed it so that it will work with the .apiRequesterData set-up, since the old "just write to GlobalEnv" absolutely does not work in a package. Plus some other cleanup, and changes needed to the initialize function to work with this.

Also added documentation.