unpooledAPIrequest won't write keys in errors
This behaviour was okay in a system where we wrote API keys to file in the repo anyway, but when we are storing the API keys in the repo secrets instead, we don't want to be leaking the keys in error messages in the tests.

Especially if we want to at some point make this repo public, we really should not ever be risking leaking our keys used for testing.